Downgrade React to v18 and update related dependencies
