--- frontend/next.config.js
@@ -1,63 +1,0 @@
-// const withNextIntl = require('next-intl/plugin')();
-
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  // Temporarily disable ESLint during build to fix production deployment
-  eslint: {
-    ignoreDuringBuilds: true,
-  },
-  // Temporarily disable TypeScript checking during build
-  typescript: {
-    ignoreBuildErrors: true,
-  },
-  // Disable static generation for problematic pages
-  output: 'standalone',
-  images: {
-    domains: ['localhost'],
-    formats: ['image/webp', 'image/avif'],
-  },
-  // PWA configuration
-  async headers() {
-    return [
-      {
-        source: '/(.*)',
-        headers: [
-          {
-            key: 'X-Frame-Options',
-            value: 'DENY',
-          },
-          {
-            key: 'X-Content-Type-Options',
-            value: 'nosniff',
-          },
-          {
-            key: 'Referrer-Policy',
-            value: 'origin-when-cross-origin',
-          },
-        ],
-      },
-    ]
-  },
-  // Optimize bundle size
-  webpack: (config, { dev, isServer }) => {
-    // Optimize images
-    config.module.rules.push({
-      test: /\.(png|jpe?g|gif|svg)$/i,
-      use: [
-        {
-          loader: 'url-loader',
-          options: {
-            limit: 8192,
-            fallback: 'file-loader',
-            publicPath: '/_next/static/images/',
-            outputPath: 'static/images/',
-          },
-        },
-      ],
-    })
-
-    return config
-  },
-}
-
-module.exports = nextConfig
